+++ Library/OcBootManagementLib/OcBootManagementLib.c
@@ -329,14 +329,11 @@ OcRunBootPicker (
         }
 
         //
-        // Do screen clearing for builtin menu here, so that it is possible to see the action.
-        // TODO: Probably remove that completely.
+        // Clear screen of previous console contents - e.g. from builtin picker,
+        // log messages or previous console tool - before loading the entry.
         //
+        gST->ConOut->ClearScreen (gST->ConOut);
         if (Context->ShowMenu == OcShowSimpleBootMenu) {
-          //
-          // Clear screen from picker contents before loading the entry.
-          //
-          gST->ConOut->ClearScreen (gST->ConOut);
           gST->ConOut->TestString (gST->ConOut, OC_CONSOLE_MARK_UNCONTROLLED);
         }